requirements.txt
pyspark==3.4.0
tweepy==4.14.0
textblob==0.17.1
pandas==2.0.3
numpy==1.24.3
hdfs3==0.3.1
pyhive==0.6.5
kafka-python==2.0.2
python-dotenv==1.0.0
fastapi==0.101.1
uvicorn==0.23.2
schedule==1.2.0
pyodbc==4.0.39
sqlalchemy==2.0.19 